Extract test_min_max_element_support.hpp.
